# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_yourorg_secret-manager_internal_models.DriftEvent:
    properties:
      created_at:
        type: string
      detected_at:
        type: string
      diff:
        items:
          type: integer
        type: array
      git_version:
        items:
          type: integer
        type: array
      id:
        type: string
      k8s_version:
        items:
          type: integer
        type: array
      namespace:
        allOf:
        - $ref: '#/definitions/github_com_yourorg_secret-manager_internal_models.Namespace'
        description: Relationships
      namespace_id:
        type: string
      resolution_action:
        type: string
      resolved_at:
        type: string
      resolved_by:
        type: string
      resolver:
        $ref: '#/definitions/github_com_yourorg_secret-manager_internal_models.User'
      secret_name:
        type: string
    type: object
  github_com_yourorg_secret-manager_internal_models.Group:
    properties:
      azure_ad_gid:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/github_com_yourorg_secret-manager_internal_models.GroupPermission'
        type: array
      users:
        description: Relationships
        items:
          $ref: '#/definitions/github_com_yourorg_secret-manager_internal_models.User'
        type: array
    type: object
  github_com_yourorg_secret-manager_internal_models.GroupPermission:
    properties:
      created_at:
        type: string
      group:
        allOf:
        - $ref: '#/definitions/github_com_yourorg_secret-manager_internal_models.Group'
        description: Relationships
      group_id:
        type: string
      id:
        type: string
      namespace:
        $ref: '#/definitions/github_com_yourorg_secret-manager_internal_models.Namespace'
      namespace_id:
        type: string
      role:
        type: string
    type: object
  github_com_yourorg_secret-manager_internal_models.Namespace:
    properties:
      cluster:
        type: string
      created_at:
        type: string
      drift_events:
        items:
          $ref: '#/definitions/github_com_yourorg_secret-manager_internal_models.DriftEvent'
        type: array
      environment:
        type: string
      id:
        type: string
      name:
        type: string
      permissions:
        description: Relationships
        items:
          $ref: '#/definitions/github_com_yourorg_secret-manager_internal_models.GroupPermission'
        type: array
      secret_drafts:
        items:
          $ref: '#/definitions/github_com_yourorg_secret-manager_internal_models.SecretDraft'
        type: array
    type: object
  github_com_yourorg_secret-manager_internal_models.SecretDraft:
    properties:
      commit_sha:
        type: string
      created_at:
        type: string
      data:
        items:
          type: integer
        type: array
      edited_at:
        type: string
      edited_by:
        type: string
      editor:
        $ref: '#/definitions/github_com_yourorg_secret-manager_internal_models.User'
      git_base_sha:
        type: string
      id:
        type: string
      namespace:
        allOf:
        - $ref: '#/definitions/github_com_yourorg_secret-manager_internal_models.Namespace'
        description: Relationships
      namespace_id:
        type: string
      published_at:
        type: string
      published_by:
        type: string
      publisher:
        $ref: '#/definitions/github_com_yourorg_secret-manager_internal_models.User'
      secret_name:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  github_com_yourorg_secret-manager_internal_models.User:
    properties:
      azure_ad_oid:
        type: string
      created_at:
        type: string
      email:
        type: string
      groups:
        description: Relationships
        items:
          $ref: '#/definitions/github_com_yourorg_secret-manager_internal_models.Group'
        type: array
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/yourorg/secret-manager/issues
  description: Production-ready secret management with drift detection and GitOps
    integration
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/yourorg/secret-manager
  title: Secret Manager API
  version: "1.0"
paths:
...
  /namespaces/{namespace}/secrets/{name}/publish:
    post:
      consumes:
      - application/json
      description: Encrypts secret with SOPS and commits to Git repository for GitOps
        sync to Kubernetes
      parameters:
      - description: Namespace ID (UUID)
        in: path
        name: namespace
        required: true
        type: string
      - description: Secret name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Published secret with commit SHA
          schema:
            $ref: '#/definitions/github_com_yourorg_secret-manager_internal_models.SecretDraft'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Namespace or secret not found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Secret has unresolved drift
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Git or SOPS not configured
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Publish secret to Git
      tags:
      - secrets
  /namespaces/{namespace}/secrets/{name}/unpublish:
    post:
      consumes:
      - application/json
      description: Removes secret from Git repository and reverts status to draft
      parameters:
      - description: Namespace ID (UUID)
        in: path
        name: namespace
        required: true
        type: string
      - description: Secret name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Unpublished secret
          schema:
            $ref: '#/definitions/github_com_yourorg_secret-manager_internal_models.SecretDraft'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Namespace or secret not found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Secret is not published
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Unpublish secret from Git
      tags:
      - secrets
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"